--- css/forget.css
@@ -1,85 +1,0 @@
-.forget-tab{
-    display: flex;
-    align-content: flex-start;
-    justify-content: center;
-    
-}
-.forget-container{
-    align-items: center;
-    flex-direction: row;
-}
-.forget-container p {
-    top: 1rem;
-}
-    
-#forget-username{
-    margin: 1rem 0 1rem 0;
-}
-
-#forget-section{
-    margin-bottom:1rem ;
-    color:var(--text-primary);
-}
-.forget-label, .labs{
-    position: absolute;
-    transition: all 0.8s ease;
-     
-    font-size: 13px;
-    color: var(--text-secondary);
-}
-.forget-input:focus + .forget-label,
-.forget-input:not(:placeholder-shown) + .forget-label{
-transform: translateX(5rem);
-opacity: 0;
-}
-.frgt-form{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-.submit-btn{
-    margin: .5rem;
-    padding: 0.5rem;
-    width: 10.8rem;
-    background:var(--btn-bg);
-    backdrop-filter: blur(10px);
-    -webkit-backdrop-filter: blur(10px);
-    border: solid 0.5px rgba(250, 250, 250, 0.322);
-    box-shadow: 2px 4px 20px rgba( 0,0,0,0.5);
-    border-radius: 12px;
-    outline: none;
-    color: var(--text-primary);
-    transition: all 0.5s ease;
-}
-.log-back-btn{
-    margin: .5rem;
-    background: transparent;
-    border-radius: 12px;
-    outline: none;
-    border: none;
-    color: var(--background-dot);
-    position: relative;
-    bottom: -2rem;
-}
- .frgt-image{
-    max-width: 100%;
-    display: block;
-    position: relative;
-    height: 50px;
-    width: 50px;
-    left: 4.5rem;
-    margin-bottom: 1rem;
-} 
-.mail-icon{
-    position: absolute;
-    height: 20px;
-    width: 20px;
-    right: 0rem;
-    bottom:1.5rem;
-}
-    
-/* .forget-password-form{
-    display: flex;
-    position: relative;
-    flex-direction: column;
-} */
