--- css/layout.css
@@ -1,50 +1,0 @@
-:root {
-  /*  Podstawowe kolory */
-  --color-bg: #0d0d0d; /* Tło główne */
-  --color-bg-soft: #1c1c1e; /* Karty, modale */
-  --color-glass-bg: rgba(22, 22, 22, 0.6); /* Glassmorphic okna */
-  --color-text-primary: #ffffff;
-  --color-text-secondary: #aaaaaa;
-  --color-divider: rgba(117, 112, 112, 0.404);
-
-  /*  Akcenty */
-  --color-accent: #6bcdce;
-  --color-accent-light: #9be5ec;
-  --color-accent-border: #44787c;
-
-  --color-accent-blur: #9be5ec8e;
-  --color-border-task: rgba(107, 106, 106, 0.1);
-  --color-subtask: #262629;
-  --color-subtask2: #2f394da2;
-
-  /*  Statusy */
-  --color-done: #5ef3a3;
-  --color-danger: #ff5f5f;
-  --color-warning: #ffc857;
-
-  /*  Elementy UI */
-  --play-btn-light: #9be5ec;
-  --play-btn-border: #44787c;
-  --play-btn-bg: rgba(107, 205, 206, 0.1);
-
-  --continue-btn-light: #62c76b;
-  --continue-btn-border: #4fa859;
-  --continue-btn-bg: rgba(98, 199, 107, 0.1);
-
-  --pause-btn-main: #e3b341;
-  --pause-btn-border: #d6a835;
-  --pause-btn-background: rgba(227, 179, 65, 0.1);
-
-  --color-red-btn: #db5656;
-  --color-text-small: #333;
-  --color-text-primiary-contrast: #1d2b36;
-  --color-text-secondary-contrast: #2b2b2b;
-  --color-hamburger: #1c1c1e96;
-  --color-button-hover: rgba(255, 255, 255, 0.05);
-  --color-input-border: rgba(255, 255, 255, 0.15);
-  --color-scrollbar: rgba(255, 255, 255, 0.1);
-  --color-btns: #2a2a2c;
-  --color-border-medium: #ffa857;
-  --color-border-height: #f53434;
-  --box-shadow-white: rgba(255, 255, 255, 0.219);
-}
