--- css/login.css
@@ -1,185 +1,0 @@
-:root {
-  --blacky: rgba(15, 29, 43, 0.8);
-  --fonty: rgba(180, 210, 230, 0.9);
-  --darky: #0d1b2a;
-  --lighty: #1b263b;
-  /* GLASS */
-  --glass-bg: rgba(255, 255, 255, 0.05);
-  --glass-border: rgba(255, 255, 255, 0.1);
-  --glass-shadow: rgba(0, 0, 0, 0.3);
-  --background-dot: #6bcdce;
-  /* sekcje */
-  --sction-bg: #212224;
-
-  /* TEKST */
-  --text-primary: #f8fafc;
-  --text-secondary: #cbd5e1;
-  --text-muted: #94a3b8;
-  --text-inverted: #0f172a;
-
-  /* PRZYCISKI */
-  --btn-bg: rgba(255, 255, 255, 0.08);
-  --btn-hover: rgba(255, 255, 255, 0.15);
-  --btn-accent: #38bdf8;
-  --btn-danger: #ef4444;
-  --btn-possitive: #0ea70e;
-  --inside-btn-bg: rgba(17, 39, 41, 0.08);
-}
-
-/* .screen {
-  display: flex;
-  position: relative;
-  flex-wrap: wrap;
-  justify-content: center;
-  width: 22rem;
-  min-height: 22rem;
-  border-radius: 12px;
-  background-color: var(--glass-border);
-  backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
-  text-align: center;
-  border: solid 0.5px var(--glass-border);
-} */
-
-/* .login {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 20rem;
-  overflow: hidden;
-
-  position: relative;
-  color: var(--text-muted);
-} */
-/* input {
-  position: relative;
-  background: transparent;
-  border: none;
-  border-bottom: 0.5px solid var(--text-muted);
-  outline: none;
-  width: 10.8rem;
-  line-height: 1.5rem;
-  color: var(--color-text-secondary);
-}
-.input-error {
-  border-bottom: 0.5px solid var(--btn-danger);
-}
-
-.input-possitive {
-  border-bottom: 0.5px solid var(--btn-possitive);
-}
-.labs:hover {
-  transform: scale(1.1);
-}
-
-.log-label,
-.pass-label {
-  position: absolute;
-  transition: all 0.8s ease;
-
-  font-size: 13px;
-  color: var(--text-secondary);
-}
-.log-input:focus + .log-label,
-.log-input:not(:placeholder-shown) + .log-label {
-  transform: translateX(5rem);
-  opacity: 0;
-}
-.pass-input:focus + .pass-label,
-.pass-input:not(:placeholder-shown) + .pass-label {
-  transform: translateX(5rem);
-  opacity: 0;
-}
-.pass-input {
-  margin-bottom: 0.5rem;
-} */
-
-/* .login-btn {
-  margin: 0.5rem;
-  padding: 0.5rem;
-  width: 10.8rem;
-  background: var(--btn-bg);
-  backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
-  border: solid 0.5px rgba(250, 250, 250, 0.322);
-  box-shadow: 2px 4px 20px rgba(0, 0, 0, 0.5);
-  border-radius: 12px;
-  outline: none;
-  color: var(--text-primary);
-  transition: all 0.5s ease;
-} */
-/* .btn:hover {
-  transform: scaleX(1.07);
-  text-decoration: underline;
-} */
-/* .log-q {
-  font-size: 12px;
-}
-.forget-link {
-  margin: 0.5rem;
-  background: transparent;
-  border-radius: 12px;
-  outline: none;
-  border: none;
-  color: var(--background-dot);
-}
-.register-link {
-  margin-left: 1.6rem;
-  background: transparent;
-  border-radius: 12px;
-  outline: none;
-  border: none;
-  color: var(--background-dot);
-} */
-.input-container {
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  position: relative;
-}
-#username {
-  margin-bottom: 0.7rem;
-}
-.forget-password,
-.register {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  color: var(--text-muted);
-}
-#login-section {
-  margin-bottom: 1rem;
-  color: var(--text-primary);
-}
-/* .login-form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-} */
-.register {
-  padding-bottom: 0.8rem;
-}
-h1 {
-  display: flex;
-  color: var(--text-primary);
-}
-header {
-  position: relative;
-  display: flex;
-  bottom: 11rem;
-}
-footer {
-  font-size: 12px;
-  position: absolute;
-  top: 32rem;
-  color: var(--text-muted);
-}
-.user-image {
-  max-width: 100%;
-  display: block;
-  position: relative;
-  height: 50px;
-  width: 50px;
-  margin-bottom: 0.5rem;
-}
