--- css/register.css
@@ -1,53 +1,0 @@
-.register-tab{
-    display: flex;
-    flex-direction: column;
-    align-content: center;
-    justify-content: center;
-    align-items: center;
-}
-.reg-form{
-    display: flex;
-    
-    flex-direction: column;
-}
-.reg-btn{
-    margin: 1rem 0 0 0;
-    padding: 0.5rem;
-    width: 10.8rem;
-    background:var(--btn-bg);
-    backdrop-filter: blur(10px);
-    -webkit-backdrop-filter: blur(10px);
-    border: solid 0.5px rgba(250, 250, 250, 0.322);
-    box-shadow: 2px 4px 20px rgba( 0,0,0,0.5);
-    border-radius: 12px;
-    outline: none;
-    color: var(--text-primary);
-    transition: all 0.5s ease;
-}
-.reg-head{
-    margin-bottom:.5rem ;
-    color:var(--text-primary);
-}
-.reg-inputs{
-    line-height: 1.5rem;
-    margin-bottom: .3rem;
-}
-.reg-back-btn{
-    position: relative;
-    top: .5rem;
-}
-
-.reg-inputs:focus + .labs,
-.reg-inputs:not(:placeholder-shown) + .labs{
-transform: translateX(5rem);
-opacity: 0;
-}
-.reg-book{
-    max-width: 100%;
-    display: block;
-    position: relative;
-    height: 50px;
-    width: 50px;
-    left: 4.2rem;
-    margin-bottom: .5rem;
- }
