--- css/roots.css
@@ -1,61 +1,0 @@
-:root {
-  /* TŁA */
-  --bg-gradient: linear-gradient(135deg, #0F172A, #1E293B); /* ciemny granat z odcieniem granatowo-niebieskim */
-  --bg-solid: #0F172A; /* fallback */
-
-  /* GLASS */
-  --glass-bg: rgba(255, 255, 255, 0.05);
-  --glass-border: rgba(255, 255, 255, 0.1);
-  --glass-shadow: rgba(0, 0, 0, 0.3);
-  /* sekcje */
-  --sction-bg: #2d2e30;
-
-  /* TEKST */
-  --text-primary: #F8FAFC;
-  --text-secondary: #CBD5E1;
-  --text-muted: #94A3B8;
-  --text-inverted: #0F172A;
-
-  /* PRZYCISKI */
-  --btn-bg: rgba(255, 255, 255, 0.08);
-  --btn-hover: rgba(255, 255, 255, 0.15);
-  --btn-accent: #38BDF8;
-  --btn-danger: #EF4444;
-
-  /* AKCENTY */
-  --accent-blue: #38BDF8;
-  --accent-cyan: #22D3EE;
-  --accent-yellow: #FACC15;
-  --accent-red: #EF4444;
-
-  /* NOISE */
-  --noise-opacity: 0.015;
-}
-
-:root {
-  /*  Podstawowe kolory */
-  --color-bg: #0d0d0d; /* Tło główne */
-  --color-bg-soft: #161616; /* Karty, modale */
-  --color-glass-bg: rgba(22, 22, 22, 0.6); /* Glassmorphic okna */
-  --color-text-primary: #ffffff;
-  --color-text-secondary: #aaaaaa;
-  --color-divider: rgba(255, 255, 255, 0.07);
-
-  /*  Akcenty */
-  --color-accent: #6BCDCE;
-  --color-accent-light: #9BE5EC;
-
-  /*  Statusy */
-  --color-done: #5ef3a3;
-  --color-danger: #ff5f5f;
-  --color-warning: #ffc857;
-
-  /*  Elementy UI */
-  --color-button-hover: rgba(255, 255, 255, 0.05);
-  --color-input-border: rgba(255, 255, 255, 0.15);
-  --color-scrollbar: rgba(255, 255, 255, 0.1);
-
-  /*  Efekt glass */
-  --glass-blur: blur(16px);
-  --glass-radius: 12px;
-}
