--- css/utils.css
@@ -1,21 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Oxanium:wght@200..800&display=swap");
-body {
-  font-family: "Oxanium", sans-serif;
-}
-
-.hidden {
-  display: none !important;
-}
-.open {
-  display: block;
-}
-.visually-hidden {
-  position: absolute !important;
-  width: 1px !important;
-  height: 1px !important;
-  margin: -1px !important;
-  padding: 0 !important;
-  overflow: hidden !important;
-  clip: rect(0, 0, 0, 0) !important;
-  border: 0 !important;
-}
