+++ index.html
@@ -5,16 +5,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
     <link rel="stylesheet" href="css/reset.css">
-    <link rel="stylesheet" href="css/layout.css">
-    
-    <link rel="stylesheet" href="css/buttons.css">
-    <link rel="stylesheet" href="css/login.css">
-    <link rel="stylesheet" href="css/register.css">
-    <link rel="stylesheet" href="css/forget.css">
-    
     <link rel="stylesheet" href="css/animations.css">
-    <link rel="stylesheet" href="css/utils.css">
     <link rel="stylesheet" href="css/main.css">
+    
+    
+    
+    
+    
+    
     <!-- LIBRARIES: -->
         <!-- Animate.css: -->
          <link
@@ -52,10 +50,24 @@
     <main class ="auth__main">
          
         <section class="login auth__pane" aria-labelledby="login-section">
-            <img  src="assets/user-shield-alt-1-svgrepo-com.svg"
-                  class="user-image"
-                  alt=""
-                  aria-hidden="true"/>
+            <span class="auth__icon" 
+                  aria-hidden="true">
+                  <svg xmlns="http://www.w3.org/2000/svg"
+                       fill="none"
+                       viewBox="0 0 24 24"
+                       stroke-width="1.5"
+                       stroke="currentColor"
+                       
+                       focusable="false">
+                       <path stroke-linecap="round"
+                             stroke-linejoin="round"
+                             d="M18 18.72a9.094 9.094 0 0 0 3.741-.479 3 3 0 0 0-4.682-2.72m.94 3.198.001.031c0 .225-.012.447-.037.666A11.944 11.944 0 0 1 12 21c-2.17 0-4.207-.576-5.963-1.584A6.062 6.062 0 0 1 6 18.719m12 0a5.971 5.971 0 0 0-.941-3.197m0 0A5.995 5.995 0 0 0 12 12.75a5.995 5.995 0 0 0-5.058 2.772m0 0a3 3 0 0 0-4.681 2.72 8.986 8.986 0 0 0 3.74.477m.94-3.197a5.971 5.971 0 0 0-.94 3.197M15 6.75a3 3 0 1 1-6 0 3 3 0 0 1 6 0Zm6 3a2.25 2.25 0 1 1-4.5 0 2.25 2.25 0 0 1 4.5 0Zm-13.5 0a2.25 2.25 0 1 1-4.5 0 2.25 2.25 0 0 1 4.5 0Z" />
+                       
+                    </svg>
+            </span>  
+                  
+
+           
 
             <h2 id="login-section"
                 class="auth-form__title">Sign in
@@ -768,10 +780,20 @@
         <section class="register-form auth__pane"
                  aria-labelledby="register-section">
 
-            <img  src="assets/user-plus-alt-1-svgrepo-com.svg" 
-                  class="reg-book" 
-                  alt=""
-                  aria-hidden="true"/>
+            <span class="auth__icon" 
+                  aria-hidden="true">
+                  <svg xmlns="http://www.w3.org/2000/svg"
+                       fill="none"
+                       viewBox="0 0 24 24"
+                       stroke-width="1.5"
+                       stroke="currentColor"
+                       class="size-6">
+                       <path stroke-linecap="round"
+                             stroke-linejoin="round"
+                             d="M18 7.5v3m0 0v3m0-3h3m-3 0h-3m-2.25-4.125a3.375 3.375 0 1 1-6.75 0 3.375 3.375 0 0 1 6.75 0ZM3 19.235v-.11a6.375 6.375 0 0 1 12.75 0v.109A12.318 12.318 0 0 1 9.374 21c-2.331 0-4.512-.645-6.374-1.766Z" />
+                    </svg>
+            </span>
+
             <h2 id = "register-section" 
                 class="reg-head auth-form__title">
                 Create your account
@@ -859,10 +881,24 @@
     <main class="auth__main">
         <section class="forget-password-form auth__pane"
                  aria-labelledby="forget-section">
-            <img  src="assets/forgot-password.svg"
-                  class="frgt-image"
-                  alt=""
-                  aria-hidden="true"/>
+           <span class="auth__icon" 
+                  aria-hidden="true">
+                  <svg fill="currentColor"
+                       focusable="false"
+                       viewBox="0 0 512 512"
+                       xmlns="http://www.w3.org/2000/svg">
+                       <g id="SVGRepo_bgCarrier"
+                          stroke-width="0"></g>
+                       <g id="SVGRepo_tracerCarrier"
+                          stroke-linecap="round" 
+                          stroke-linejoin="round"></g>
+                       <g id="SVGRepo_iconCarrier">
+                       <g id="Change_password">
+                       <path d="M464.4326,147.54a9.8985,9.8985,0,0,0-17.56,9.1406,214.2638,214.2638,0,0,1-38.7686,251.42c-83.8564,83.8476-220.3154,83.874-304.207-.0088a9.8957,9.8957,0,0,0-16.8926,7.0049v56.9a9.8965,9.8965,0,0,0,19.793,0v-34.55A234.9509,234.9509,0,0,0,464.4326,147.54Z"></path>
+                       <path d="M103.8965,103.9022c83.8828-83.874,220.3418-83.8652,304.207-.0088a9.8906,9.8906,0,0,0,16.8926-6.9961v-56.9a9.8965,9.8965,0,0,0-19.793,0v34.55C313.0234-1.3556,176.0547,3.7509,89.9043,89.9012A233.9561,233.9561,0,0,0,47.5674,364.454a9.8985,9.8985,0,0,0,17.56-9.1406A214.2485,214.2485,0,0,1,103.8965,103.9022Z"></path>
+                       <path d="M126.4009,254.5555v109.44a27.08,27.08,0,0,0,27,27H358.5991a27.077,27.077,0,0,0,27-27v-109.44a27.0777,27.0777,0,0,0-27-27H153.4009A27.0805,27.0805,0,0,0,126.4009,254.5555ZM328,288.13a21.1465,21.1465,0,1,1-21.1465,21.1464A21.1667,21.1667,0,0,1,328,288.13Zm-72,0a21.1465,21.1465,0,1,1-21.1465,21.1464A21.1667,21.1667,0,0,1,256,288.13Zm-72,0a21.1465,21.1465,0,1,1-21.1465,21.1464A21.1667,21.1667,0,0,1,184,288.13Z"></path>
+                       <path d="M343.6533,207.756V171.7538a87.6533,87.6533,0,0,0-175.3066,0V207.756H188.14V171.7538a67.86,67.86,0,0,1,135.7208,0V207.756Z"></path> </g> </g></svg>
+            </span>
             <h2 id="forget-section" 
                 class="auth-form__title">
                 Reset your password
@@ -874,11 +910,7 @@
                   method="POST"
                   novalidate>
 
-                <p class="auth-form__status hidden"
-                   role="status"
-                   aria-live="polite">
-                   If the email exists, we'll send a password reset link.
-                </p>
+                
 
                 <div class="input-container forget-container auth-form__field">
                     <p class="err-message hidden auth-form__error"
