+++ js/Services/view-mangers/carousel-settings.js
