+++ js/components/login-reg-forget/authUIController.js
@@ -1,7 +1,7 @@
 import { LoginFormHandler } from './formHandlers.js';
-import { FormErrors } from './uiErrorHandler.js';
+import { FormErrors } from '../../Services/validators/uiErrorHandler.js';
 import { getAuth } from 'https://www.gstatic.com/firebasejs/11.8.1/firebase-auth.js';
-import { fireApp } from './firebase-init.js';
+import { fireApp } from '../../Services/firebase/firebase-init.js';
 
 export class AuthUIController {
   constructor(viewManager) {
@@ -13,51 +13,44 @@ export class AuthUIController {
     this.getActiveView();
   }
 
-  //Metoda obsługi zmiany widoków
   bindScreenEvents() {
-    // Pobieram przyciski zmiany widokow:
     const logBackBtns = document.querySelectorAll('.log-back-btn'); //<- wraca do logowania
     const forgetBtn = document.querySelector('.forget-link'); //<- idzie do sekcji odzyskaj haslo
     const regstrBtn = document.querySelector('.register-link'); //<- idzie do sekcji rejestracji
 
-    // dodaje nasłuchiwanie na przycisk "zapomniałes Hasło"
     forgetBtn.addEventListener('click', (e) => this.showFrgt(e));
 
-    // dodaje Nasłuchiwanie na przycisk Rejsetracji
     regstrBtn.addEventListener('click', (e) => this.showRgstr(e));
 
-    //dodaje nasluchiwanie na przycisk poworotu do logowania
     logBackBtns.forEach((btns) =>
       btns.addEventListener('click', (e) => this.showLogIn(e))
     );
   }
-  // Metoda pokazujaca forget
+
   showFrgt(e) {
     e.preventDefault();
-    console.log('kliknieto w forget');
+
     this.viewManager.showView('forget-screen');
   }
-  // Metoda pokazująca  rejstracje
+
   showRgstr(e) {
     e.preventDefault();
-    console.log('kliknieto w register');
+
     this.viewManager.showView('register-screen');
   }
-  // Metoda powrotu do Logowania
+
   showLogIn(e) {
     e.preventDefault();
     this.viewManager.hideOneView('forget-screen');
-    console.log('kliknieto w logback');
+
     this.viewManager.showView('login-screen');
   }
-  //metoda do zmiany widoku po wylogowaniu sie
-  logout() {
-    console.log('wylogowano');
 
+  logout() {
     this.viewManager.hideOneView('todo-screen');
     this.viewManager.showView('login-screen');
   }
-  //sprawdzamy i okreslamy aktualny widok
+
   getActiveView() {
     if (document.getElementById('login-screen')?.offsetParent !== null) {
       return 'login';
